Gym-Recommendation-and-Training-Split-Management-System-main/gym_ai_service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic>=2.9.0
python-dotenv==1.0.1
mistralai==1.5.0
